ハンドラーをmainからパッケージに切り出す。
