+++ app/template/parts/header.html
@@ -25,7 +25,7 @@
   <nav class="navbar navbar-default mb-3">
     <div class="container-fluid">
       <div class="navbar-header">
-        <a class="navbar-brand" href="/book">余田本リスト</a>
+        <a class="navbar-brand" href="/book">みんなの本棚</a>
       </div>
       <div class="">
         <a href="/book/regist" class="btn btn-primary text-white">本を登録する</a>
